requirements.txt
Django==1.8.2
djangorestframework==3.1.3
django-enumfields==0.7.3
